ai-business-intelligence/ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.15.0
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
requests==2.31.0
